Improve course-editing behaviour
Previously when creating a new course, typing into the course code box,
then clearing it out would result in a course being saved containing
the first letter of the code that was typed in.

A new AutoSavingItem type was created to facilitate the creation of a
course. This allows classes to revert a save they previously committed.